refactor and exception handling for dns_cli.py
