+++ .gitignore
@@ -1,3 +1,4 @@
 .idea/
-
+.vscode/
 venv/
+.pytest_cache/
