+++ requirements.txt
@@ -6,3 +6,5 @@ PyYAML==6.0
 requests==2.31.0
 Jinja2==3.1.2
 schema==0.7.5
+pytest
+moto
